Added the functionality for the select
Everytime the user changes the value (city timezone) of the select it calls the function called updateCity: which is going to receive the timezone and then create a new city name based on the timezone so it's going to replace the underscore and the and use space split the name into two to second value then to New York only. then it is calculating the city timezone of the city with moment.tz from the current local time that we got from the dropdown bar. And then I am selecting the cities which are to be replaced with the whole HTML which includes the city name, the date and the current time and also the AM/PM.